+++ app/Http/Resources/Api/V1/WriterResource.php
@@ -5,15 +5,13 @@ declare(strict_types=1);
 namespace App\Http\Resources\Api\V1;
 
 use App\Models\Writer;
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
-use Illuminate\Http\Resources\Json\JsonResource;
 use Override;
 
 /**
  * @mixin Writer
  */
-final class WriterResource extends JsonResource
+final class WriterResource extends ApiResource
 {
     /**
      * @return array<string, mixed>
@@ -21,9 +19,6 @@ final class WriterResource extends JsonResource
     #[Override]
     public function toArray(Request $request): array
     {
-        /** @var Model $model */
-        $model = $this->resource;
-
-        return $model->attributesToArray();
+        return $this->normalisedAttributes();
     }
 }